remove question and onchangeQuestion from Context
