+++ src/components/comment/index.tsx
@@ -1,5 +1,4 @@
-import { useEffect, useContext, useState } from "react";
-import { ContextApplication } from "@/context";
+import { useEffect, useState } from "react";
 import TextArea from "@/components/text-area";
 import Validation from "@/components/validation";
 import { Button, BackLink } from "@/components/button";
@@ -7,9 +6,15 @@ import { questions } from "../../../util/questionsInfo";
 import { getLocalStorage } from "../../../util/helpLocalStorage";
 import { CommentForm } from "../../../util/type";
 
-function CommentQuestion() {
-  const { onChangeQuestion, setQuestion, question } =
-    useContext(ContextApplication);
+function CommentQuestion({
+  onChangeQuestion,
+  setQuestion,
+  question,
+}: {
+  onChangeQuestion: () => void;
+  setQuestion: (value: number) => void;
+  question: number;
+}) {
   const [selectedCheckbox, setSelectedCheckbox] = useState<number[]>([]);
   const [commentForm, setCommentForm] = useState<CommentForm>({});
   const [isError, setIsError] = useState(false);
