+++ src/context/helper.tsx
@@ -1,25 +1,4 @@
 export const defaultValue = {
-  dataApplication: {},
-  setDataApplication: () => {},
-  question: 0,
-  setQuestion: () => {},
-  selectedCheckbox: [],
-  setSelectedCheckbox: () => {},
-  onChangeQuestion: () => {},
-  feelingForm: "",
-  setFeelingForm: () => {},
-  commentForm: {},
-  setCommentForm: () => {},
-  personalDetailsForm: {
-    name: "",
-    address: "",
-    email: "",
-    phone: "",
-    postcode: "",
-    consent: false,
-  },
-  setPersonalDetailsForm: () => {},
   globalInfo: {},
   setGlobalInfo: () => {},
-  contextCleaner: () => {},
 };
