+++ src/pages/planning-applications/[id]/feedback/index.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState, useContext } from "react";
-import { ContextApplication } from "@/context";
 import Breadcrumbs from "@/components/breadcrumbs";
 import Instructions from "./instructions";
 import Questions from "../../../../components/questions";
@@ -7,8 +6,9 @@ import { getLocalStorage } from "../../../../../util/helpLocalStorage";
 import { Data } from "../../../../../util/type";
 
 const Feedback = () => {
-  const { question } = useContext(ContextApplication);
   const [application, setApplication] = useState<Data>();
+  const [question, setQuestion] = useState<number>(0);
+  const [selectedCheckbox, setSelectedCheckbox] = useState<number[]>([]);
 
   useEffect(() => {
     const getStorage = getLocalStorage({
@@ -16,7 +16,35 @@ const Feedback = () => {
       defaultValue: {},
     });
     setApplication(getStorage);
-  }, []);
+
+    const getStorageSelectedCheckbox = getLocalStorage({
+      key: "impact",
+      defaultValue: selectedCheckbox,
+    });
+    getStorage?._id === getStorageSelectedCheckbox?.id &&
+      setSelectedCheckbox(getStorageSelectedCheckbox?.value);
+  }, [selectedCheckbox]);
+
+  const onChangeQuestion = () => {
+    if (
+      question === 0 ||
+      question === 1 ||
+      question === 11 ||
+      question === 12
+    ) {
+      setQuestion(question + 1);
+    } else if (question === 2) {
+      setQuestion(selectedCheckbox[0]);
+    } else if (question === 13) {
+      setQuestion(0);
+    } else if (question === selectedCheckbox[selectedCheckbox.length - 1]) {
+      setQuestion(11);
+    } else {
+      const questionIndex = selectedCheckbox?.indexOf(question);
+      const newQuestion = selectedCheckbox[questionIndex + 1];
+      setQuestion(newQuestion);
+    }
+  };
 
   const breadcrumbs_array = [
     {
@@ -36,7 +64,11 @@ const Feedback = () => {
     <>
       <Breadcrumbs breadcrumbs_info={breadcrumbs_array} />
       {question !== 0 && question !== 13 && <Instructions />}
-      <Questions question={question} />
+      <Questions
+        question={question}
+        onChangeQuestion={() => onChangeQuestion()}
+        setQuestion={() => setQuestion}
+      />
     </>
   );
 };
